tests: improve portability under qemu/binfmt
When building with cross-compiling AND running (non-native) binary
locally with qemu/binfmt, the argv[0] (binary name) is changed by qemu.
argv[0] is used for error reporting.

Example with most systems, argv[0] is exactly the same as the user
entered:
    $ datamash --foobar
    datamash: invalid option --foobar

    $ /custom/path/datamash --foobar
    /custom/path/datamash: invalid option --foobar

But under qemu/binfmt, argv[0] becomes the full path:

    $ datamash --foobar
    /home/gordon/projects/datamash/datamash: invalid option --foobar

The tests were modified to detect the actual path reported by the
program, or fall back to 'datamash' (as it was before).