datamash: keep correct lines with '--full'
Discussed in https://github.com/agordon/datamash/issues/3 ,
and also existed in (very) old revisions of 'calc'.

When using a selection operation (first/last/min/max/etc) which
selects one line out of the group, and combining with --full,
this ensures the correct line is selected (not just the correct
fields intemixed with the values of the first line).

* src/datamash.c: keep the correct line
* src/field-ops.{c,h}: for selection operation, determine whether to
    keep the line or not.
* src/datamash-tests.pl: adapt one test of 'lst' to new behavior.
* src/datamash-tests-2.pl: tests new --full behavour