datamash: new operation: cut
Similar to cut(1), it copies the input field to the output as-is.
The advantage over cut(1) is that combined with datamash's other features,
input fields can be specified by name instead of column number,
and output fields can be re-ordered and duplicated.

  Example:
    $ printf "a b c\n1 X 6\n" | datamash -W -H cut c,a,c
    cut(c)  cut(a)  cut(c)
    6       1       6

Suggested by Torsten Seemann <torsten.seemann@gmail.com> in
https://lists.gnu.org/archive/html/bug-datamash/2019-09/msg00000.html

* NEWS: Mention new operation.
* src/datamash.c (usage): Add new op.
* src/op-defs.{h,c}: Define new OP_CUT.
* src/field-ops.{c}: Implement new OP_CUT.
* contrib/bash-completion/datamash: Add new operation.
* doc/datamash.texi, man/datamash.x: Mention new operation.
* tests/datamash-tests-2.pl: Add tests.