decorate: new conversions ipv6v4map and ipv6v4comp
The two new decorate(1) conversion methods ipv6v4map and
ipv6v4comp allow to sort on a field that may contain either
an IPv6 or an IPv4 address.

IPv4 and IPv6 addresses can be seen as IP addresses.  Logs from
a "dual-stack" application, e.g., a web server, may contain
either an IPv4 or IPv6 address at a given position in each line.
Thus if one wants to sort the log file on the IP address,
both IPv4 and IPv6 addresses need to be accepted as sort key
and sorted consistently.  One approach is to transform one
address type into the other before sorting.  IPv6 supports
the transformation of IPv4 addresses into IPv6 addresses.

There are two common methods for accomodating IPv4 addresses in
IPv6: IPv4-Mapped addresses and the deprecated IPv4-Compatible
addresses.  Both can be used to convert a given IPv4 address
to an IPv6 address.  Both IPv4-Mapped and IPv4-Compatible IPv6
address ranges are reserved by IANA and always represent IPv4
addresses in a dual-stack enabled application.

IPv4-Compatible addresses just add 96 leading zero bits to
the 32 bit IPv4 address to create a 128 bit IPv6 address.
This results in an ambiguity for the unspecified address
(all-zero in both IPv4 and IPv6) and the IPv6 localhost address
::1 with the first host address of "this" network in IPv4
(0.0.0.1).  IPv4-Mapped addresses avoid this ambiguity.

But since IPv4-Compatible IPv6 addresses can be seen as
treating the IP address (both version 4 and version 6) as
a specific way to represent an integer value I think it is
useful to support this transformation as well.

Both conversions logically convert an IPv4 address to an IPv6
address, but the code actually creates a textual representation
of an 128 bit integer from either an IPv4 or IPv6 address.

Functionality like this was requested for sort(1) from GNU
Coreutils:
  https://lists.gnu.org/archive/html/coreutils/2011-06/msg00078.html
  https://lists.gnu.org/r/bug-coreutils/2015-06/msg00039.html

It was rejected for sort(1) from GNU Coreutils:
  https://www.gnu.org/software/coreutils/rejected_requests.html#sort
  https://lists.gnu.org/r/bug-coreutils/2015-06/msg00041.html

Thus it seems appropriate for decorate(1), complementing the
existing conversion methods for either IPv4 or IPv6 addresses,
but not both.

* NEWS: Mention new decorate conversion methods.
* src/decorate-functions.c: Implement new conversion methods.
  (decorate_ipv6_ipv4): New function.
  (decorate_ipv6_ipv4_mapped): New function.
  (decorate_ipv6_ipv4_compat): New function.
  (builtin_conversions): Add new operations.
* tests/decorate-errors.pl,
  tests/decorate-sort-tests.pl,
  tests/decorate-tests.pl: Add tests for new conversion methodss.