doc: leading and trailing whitespace with -W
With the -W, --whitespace option, datamash(1) ignores leading
whitespace, but not trailing whitespace.  Trailing whitespace
results in an additional empty field at the end of the line.
This is different to field splitting in Awk, Bash, or Python.

This behavior is tested for, but not documented.  Thus mention
it in the texinfo documentation.

* doc/datamash.texi: Add a senetence to the -W, --whitespace
  option description.