doc: tweak help output and man pages
Mention the LC_NUMERIC environment variable in datamash --help
output and the generated man page.

Add an "Options:" header line to the output of datamash --help
and decorate --help so that help2man detects the OPTIONS section.

Use DESCRIPTION instead of OVERVIEW to add text from the man page
template to the DESCRIPTION section of the decorate(1) man page.

Move the examples from the decorate(1) man page template into the
EXAMPLES section of the generated man page.

* man/decorate.x: Rename OVERVIEW to DESCRIPTION.  Move EXAMPLES.
* src/datamash.c (usage): Add "Options:" header and "Environment:"
  section to help output.
* src/decorate.c (usage): Add "Options:" header to help output.