tests: enable valgrind test to pass
The sub-test "custom-format" of the expensive test
datamash-valgrind.sh would always fail, because the
datamash was supposed to sum all input numbers, but
some of those are too big for 80-bit extended floating
point numbers as used on x86.  Thus datamash would
emit an error message and return an exit code of 1.
That was interpreted as a test failure.

* tests/datamash-valgrind.sh: Use different valgrind error
  exit code to distinguish between valgrind detecting a
  memory leak and datamash reporting an error.