datamash: fix segmentation fault
As reported by Catalin Patulea on bug-datamash@gnu.org in
<https://lists.gnu.org/archive/html/bug-datamash/2020-11/msg00001.html>,
GNU Datamash could crash with a segmentation fault if the
unique or countunique operations were used with input data
containing NUL bytes.

The problem was that the field_op_get_string_ptrs() function
could create more pointers than it allocated memory for if
the input data contained NUL bytes.  The solution is to add a
check to avoid writing past the end of the "ptrs" buffer.

* NEWS: Mention bug fix.
* src/field-ops.c (field_op_get_string_ptrs): Do not write past
  the end of the "ptrs" buffer.
* tests/datamash-tests-2.pl: Add tests to verify bug fix.