tests: add cheap I/O error test
The existing tests/datamash-io-errors.sh is marked as expensive
and requires two file system images prepared to provoke I/O
errors for GNU Datamash to detect and report.  As a result it
is executed less frequently than most other tests.

The new tests/datamash-io-errors-cheap.sh requires just the
availability of the special file "/dev/full" to immediately
provoke an I/O error on output.  This test requires minimal
input data and is cheap to run.

* Makefile.am (TESTS): Add new test file.
* tests/datamash-io-errors-cheap.sh: New file with one test.