tests: avoid Perl warning in datamash-i18n-de.pl
The string comparison `$lc_de eq undef` results in up to two
warnings.  If `$lc_de` is defined, the single warning

    Use of uninitialized value in string eq at
     ./tests/datamash-i18n-de.pl line 39.

is emitted.  This is caused by comparing with `undef` in the
string equality test `eq`.

If the locale `de_DE.utf8` is not found, `$lc_de` is undefined.
If `$lc_de` is undefined, two warnings are emitted:

    Use of uninitialized value $lc_de in string eq at
     ./tests/datamash-i18n-de.pl line 39.

    Use of uninitialized value in string eq at
     ./tests/datamash-i18n-de.pl line 39.

Using `defined()` to test if `$lc_de` is defined avoids this.

* tests/datamash-i18n-de.pl: Use defined() to check if a
  variable is defined.