tests: comments and empty lines
Without any options, GNU Datamash does not support comments,
and does not treat empty lines as different from any other
line.

With -C/--skip-comments, it ignores complete lines starting
with either '#' or ';' as first non-whitespace character.

With --vnlog, it ignores empty lines, complete lines where the
first non-whitespace character is a '#' (followed by either
'!' or '#' in the vnlog prologue), and trailing comments,
i.e., the part of a non-empty, non-comment, and non-header
line started with optional whitespace and a '#' character.

* tests/datamash-check.pl: Add tests with empty and comment
  lines in the input with different options.
* tests/datamash-vnlog.pl: Add trailing comment started with
  whitespace to existing input.  Add test to verify that a
  Semicolon (';') does not start a trailing comment.  Add a
  test that the header line does not support comments.