datamash: fix "antimode" operation
Both "mode" and "antimode" operations are implemented in the same
function "mode_value()" in file "src/utils.c".  The algorithm
traverses a list of sequences of numbers and attempts to remember
the longest (for mode) or shortest (for antimode).

Before, the sequence information was updated for every new number
encountered, i.e., in general before the actual sequence length
was known.  This approach works for mode, but not for antimode.

Now, the sequence information is updated only after a sequence
has ended.  There are two ways a sequence can end: (1) a different
number is read, (2) there are no more numbers to read.  In both
cases, this new sequence information (length and value) needs
to be checked against the previous longest/shortest length and
possibly updated.

* NEWS: Mention bug fix.
* src/utils.c (mode_value): Update best sequence information only
  after end of sequence.  Move update code to new function, because
  it is now used twice.
  (update_best_seq): New function.
* tests/datamash-tests-2.pl: Activate formerly broken now working
  "antimode" tests.