datamash: bug fix for decimal sep. as field sep.
When the locale's decimal separator is used as field separator,
reading a numeric value would continue over the field separator,
resulting in an "invalid numeric value" error message.  This is
now prevented by re-trying to read the numeric value after the
field contents have been copied to a NUL-terminated buffer.

* NEWS: Mention bug fix.
* THANKS: Add bug reporters.
* src/field-ops.c (field_op_collect): If strtold reads past the
  field contents, copy field contents to temporary buffer and
  try strtold again.
* tests/datamash-i18n-de.pl: Activate two previously failing
  tests, and add a test case from the recent bug report.
* tests/datamash-tests-2.pl: Activate two previously failing
  tests.