datamash: decrease visibility of local symbols
i.e. symbols that are only used inside their translation unit and don't
have a prototype declaration in a header should be hidden with static.

the main motiviation behind such a change basically is:

1) minimize the likelihood of possible (future) symbol clashes between
   different translation units or libraries
2) speed up linking
3) eliminate noise when using `-Wmissing-prototypes`, i.e. if your local
   symbols are 'clean', then all remaining missing-prototypes warnings
   point to missing includes or obsolete functions

Note that a global symbol definition that shadows another doesn't necessarily
yield a link error, in general.

And even in the constellations where it does, for fixing, it's way more
convenient to get a compile error, instead.

Dima Kogan also added in the email discussion:

Keeping your symbols localized is general good hygiene, and has even
more upsides:

- The robot reader of the code (the compiler) gains a higher level of
  understanding: it gets knowledge of the FULL list of callers of such a
  function, and is free to make deep optimizations, like inlining the
  function

- The human reader of the code also gains a higher level of
  understanding. With local functions I also know about exactly what
  paths are possible for calling them by simply doing a text search in
  one file.

Relatedly, the linker visibility settings are controllable also (gcc
-fvisibility or attribute((visibility))), but that's important for
shared libraries, and less so for executables.