Revert "rand: Add new program"
This reverts commit 0ed0482e4fb05c99ed6fdf6d3c20864cbe61acfd.

Adding rand(1) seemed desirable from the point of view of wanting a way
to benchmark datamash(1). However, a comprehensive pseudorandom number
generation engine would requires a non-trivial amount of statistical
library support. There is currently little maintainer will for such
effort.

We can "double-revert" in future if there is interest, but for the
moment it seems cleanest to not have rand(1) enter the upcoming stable
release 1.9.